python_rag/vercel/requirements.txt
asyncpg==0.30.0
fastapi==0.115.12
loguru==0.7.3
pydantic==2.11.5
PyPDF2==3.0.1
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
uvicorn==0.34.2
async-lru==2.0.5
python-docx==1.1.2
python-multipart==0.0.20

# # Google Gemini API
google-generativeai==0.8.3

# # Remove these heavy packages:
# # sentence-transformers==4.1.0
# # transformers==4.52.4
# # torch==2.7.0
# # datasets==3.6.0
# # nltk==3.9.1
# # mangum==0.17.0